gpu_service/hf_endpoints/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
numpy>=1.26.0
soundfile>=0.12.0
librosa>=0.10.0
pyannote.audio>=3.3.0
funasr>=1.3.0
python-multipart>=0.0.9
pydub>=0.25.0
transformers>=4.45.0
accelerate>=0.34.0
sse-starlette>=1.0.0
torchvision>=0.19.0
